# compare atomic and mutex performance for incrementing one variable