```
BenchmarkIncrInt64Atomics-12    	300000000	         5.59 ns/op
BenchmarkIncrInt64Mutex-12      	100000000	        15.8 ns/op
BenchmarkSetTimeAtomics-12      	20000000	       104 ns/op
BenchmarkSetTimeMutex-12        	100000000	        14.3 ns/op

```