@lyuxuan 2017-12-19
```
BenchmarkAtomicValueStore-12        	30000000	        42.1 ns/op
BenchmarkAtomicAddInt64-12          	300000000	         5.69 ns/op
BenchmarkAtomicTimeValueStore-12    	20000000	        98.6 ns/op
```