@lyuxuan 2017-12-22
## summary
use atomic.StorePointer for best per operation execution time and best contention scenario performance.

### contention benchmark:
```
BenchmarkValueStoreWithContention/Atomic/10-12         	                 3000000	       334 ns/op
BenchmarkValueStoreWithContention/AtomicStorePointer/10-12         	10000000	       157 ns/op
BenchmarkValueStoreWithContention/Mutex/10-12                      	 1000000	      1112 ns/op
BenchmarkValueStoreWithContention/Atomic/100-12                    	  500000	      2764 ns/op
BenchmarkValueStoreWithContention/AtomicStorePointer/100-12        	 1000000	      1433 ns/op
BenchmarkValueStoreWithContention/Mutex/100-12                     	  200000	      9318 ns/op
BenchmarkValueStoreWithContention/Atomic/1000-12                   	   50000	     28546 ns/op
BenchmarkValueStoreWithContention/AtomicStorePointer/1000-12       	  100000	     14360 ns/op
BenchmarkValueStoreWithContention/Mutex/1000-12                    	   30000	     61149 ns/op
BenchmarkValueStoreWithContention/Atomic/10000-12                  	    5000	    309288 ns/op
BenchmarkValueStoreWithContention/AtomicStorePointer/10000-12      	   10000	    144596 ns/op
BenchmarkValueStoreWithContention/Mutex/10000-12                   	    5000	   4657600 ns/op
BenchmarkValueStoreWithContention/Atomic/100000-12                 	     500	   2929331 ns/op
BenchmarkValueStoreWithContention/AtomicStorePointer/100000-12     	    1000	   1468127 ns/op
BenchmarkValueStoreWithContention/Mutex/100000-12                  	      30	  39160125 ns/op
```
### per operation execution time
```
BenchmarkAtomicAddInt64-12            	300000000	         5.15 ns/op
BenchmarkAtomicTimeValueStore-12      	20000000	       113 ns/op
BenchmarkAtomic16BValueStore-12       	20000000	        69.4 ns/op
BenchmarkAtomic32BValueStore-12       	20000000	        92.7 ns/op
BenchmarkAtomicPointerStore-12        	200000000	         9.66 ns/op
BenchmarkAtomicTimePointerStore-12    	200000000	         9.93 ns/op
```